EP v3.0.98: CPU Load issue fix, and more
- Use the AGT 4 and 5 in cascade to get the 32 bit timer for the RunTime Stats.
- Create a macro DWT_CYCNT to make the code more readable
- Fix the Set Idle priority error
- Remove the duplicate line in ALRM_Handler.c file that was there from a long time
- Fix a few incorrect conditionals
- Update the name of AGT for LG to be more meaningful

+++ MTU/OS_BSP/FreeRTOS_RA6/ra_cfg/aws/FreeRTOSConfig.h
@@ -221,7 +221,7 @@
             /* Aclara Added -- Start */
             #if ( configGENERATE_RUN_TIME_STATS == 1 )
             extern void       AGT_RunTimeStatsStart( void );
-            extern uint16_t   AGT_RunTimeStatsCount( void );
+            extern uint32_t   AGT_RunTimeStatsCount( void );
             #define portCONFIGURE_TIMER_FOR_RUN_TIME_STATS() AGT_RunTimeStatsStart()
             #define portGET_RUN_TIME_COUNTER_VALUE()         AGT_RunTimeStatsCount()
             #endif
